requirements.txt

# requirements.txt

# TensorFlow para el modelo
tensorflow>=2.16.2,<2.18

# Procesamiento de audio
librosa>=0.10.0
numpy>=1.25.2
scipy>=1.12.1

# Manipulación de datos
pandas>=2.1.1

# Imagen y redimensionamiento
Pillow>=10.1.0

# Visualización (opcional si quieres gráficas en la GUI)
matplotlib>=3.8.1
seaborn>=0.12.2

# GUI (Tkinter viene con Python estándar, no requiere instalación por pip)
# tk  # No instalar por pip, solo comentar si usas Tkinter

# Otras utilidades
scikit-learn>=1.3.0

# Dependencias de construcción
setuptools
wheel
